PCHR-1872: Fix Error on T&A Calendar View
Expected default behaviour for 'get' API call to Assignments entity required it
to return an array of contacts for each assignment found, however, this was not
happening. As it turned out, Assignments entity is based on CiviCRM Case
entity, and this contact array is only included in results if a case ID is
passed as parameter on API call.  However, CiviCRM wasn't able to identify the
ID field for the Assignments entity, expecting it to be either 'id' or
'assignment_id', which never happened, since it was actually 'case_id'.

Fixed the problem by creating 'id' field in Assignment's DAO, by copying values
for 'case_id' and changing its key.