PCHR-2052: Fix Calendar for Users Without Assigned Tasks
When a user had no assigned tasks but was the source for some tasks, API calls
to obtain task lists were returning an empty array for assigned tasks, while
returning an object for list of tasks where current user was the source. This
inconsistency caused problems in front-end, making it uncapable of dealing
with an array vs an object for each call.

This happens because, although the back-end was returning an array on all
calls, calls without the sequential flag generate an associative array, which
is serialized as an object in json, since there are no associative arrays in
javascript.

Fixed by sending sequential flag = 1 for all calls to get Tasks.  This way,
the result will always be an array.