PCHR-2144: Add the "access Tasks and Assignments Files" permission
Now that the angular app for T&A documents can also be accessed through
the SSP, users with the civihr_staff roles must be able to access and
uploaded files to a document.

Previously, the required permission for this was the "access Tasks and
Assignments" permission. Unfortunately, this same permission is used to
give access to some admin pages, meaning that giving civihr_staff such
permission would also given them access to some admin areas. In order to
avoid this, a new permission was added only to the file related pages.