PCHR-2277: Don't Create Extra Revisions on Custom Field Update
The implementation of create API call to Documents entity was generating up to
two revisions per call, when the following conditions were met:

- Params array contained 'case_id'
- Params array contained a custom field

This happened because, first, Document create API generated a new revision
from the document's old data, if a value for 'case_id' was found in params
array, making any updates on the new revision after. Once the initial updates
were done, it checked if there were any custom fields being updated, and if
so, made an API call to Activity's create action to update the document's
custom fields. This call also checks for case_id and creates a new revision
if it is set in params array, or if the relationship is found in the database
for the activity.

The second API call to Activity's create action was added beacuse custom
fields were not being updated by the call to Document's BAO create method.
Now, the reason for this was in order to update custom fields, API expected
the physical column names to be used, instead of their field names.

Fixed by replacing custom field names for the column names and looking for
custom fields defined for both Documents and Activities, before starting all
updates. Also removed the second API call to Activity create action, as it is
no longer needed.