PCHR-4473: Fix flaky tests in hrcore
This fixes three issues that would make the hrcore tests fail intermittently:

- FileCache returning cached modified time. Since the class uses filemtime() to
get the modified date and that PHP caches stat data about files, the test for
the modified date would faile if run multiple times, as it return the date of
a previous run

- On 77f4cadc34c1c808222153f5de679aa3ebdf04d6, some refactoring has been done
to the Reminder class. In this refactoring, the assigneeContacts variable was not
neing declared anymore, even though it was still being used a few lines below.
Also, the statement to check if the given $contactId was an assignee ID would
throw a notice in case there was no assignee and the `assignees` item was empty

- Calls to static methods as if they were instance methods in the StaffDirectoryTest