PCHR-4473: Fix for Documents attached file issues
In CiviCRM 5.11, to improve access to attached file, the URL for file downloads
has been updated. It now contains a hash with some site specific info, which
can be used to prevent people from downloading all the files from a site
simply by manually changing the file id.

This same URL is used in T&A to allow people to download the files attached to a
Document. However, we had some custom code in place that was overriding the URL
generated by Civi and removing the generated hash, resulting in users not being
able to download files.

The reason the custom code was put in place is unknown as the commit where it was
introduced doesn't have any useful information. Looking at the code, we can see that
the only reason we were dealing with the file URL was to pass the "frontend" param
to the url() method in case the uploaded file was an image. As stated in the documentation
for the URL method, the frontend param is only applicable to Wordpress and Joomla.
Since CiviHR (the system T&A is part of) runs only on Drupal, the whole custom
code was useless, so it was removed.

The fileList() method has been refactored a bit, as there was some calls to get
additional info about the file that were not really necessary.

Finally, while inspecting the code, I realized that the route
civicrm/tasksandassignments/file/display was not being used anywhere, so I deleted it